Refactor user acceptance test logic into three files
